Deep-merge service meta with pillar data
This change performs a deep-merge between the configuration collected
from */meta/telegraf.yml and the pillar data. It allows a deployer to
override a single parameter of a Telegraf input plugin without
duplicating the whole YAML snippet in the user model. For instance:

  parameters:
    telegraf:
      agent:
        input:
          nginx:
            timeout: 10

It relies on the 'defaults.merge' function which has been added in Salt
2016.3 [1].

[1] https://github.com/saltstack/salt/commit/6868237f0bcded732ea6fe2d332

Change-Id: I6b6581b9400e7bbcb8758b7f0a90d5ecaa65e74e